move script stuff to separate file
