+++ cmd/muxt/main.go
@@ -1,13 +1,10 @@
 package main
 
 import (
-	"bytes"
 	"flag"
 	"fmt"
 	"io"
 	"os"
-
-	"rsc.io/script"
 )
 
 func main() {
@@ -37,23 +34,4 @@ func handleError(err error) int {
 		return 1
 	}
 	return 0
-}
-
-func scriptCommand() script.Cmd {
-	return script.Command(script.CmdUsage{
-		Summary: "muxt",
-		Args:    "",
-	}, func(state *script.State, args ...string) (script.WaitFunc, error) {
-		return func(state *script.State) (string, string, error) {
-			var stdout, stderr bytes.Buffer
-			err := command(state.Getwd(), args, func(s string) string {
-				e, _ := state.LookupEnv(s)
-				return e
-			}, &stdout, &stderr)
-			if err != nil {
-				stderr.WriteString(err.Error())
-			}
-			return stdout.String(), stderr.String(), err
-		}, nil
-	})
 }
