+++ cmd/muxt/script.go
@@ -1,0 +1,26 @@
+package main
+
+import (
+	"bytes"
+
+	"rsc.io/script"
+)
+
+func scriptCommand() script.Cmd {
+	return script.Command(script.CmdUsage{
+		Summary: "muxt",
+		Args:    "",
+	}, func(state *script.State, args ...string) (script.WaitFunc, error) {
+		return func(state *script.State) (string, string, error) {
+			var stdout, stderr bytes.Buffer
+			err := command(state.Getwd(), args, func(s string) string {
+				e, _ := state.LookupEnv(s)
+				return e
+			}, &stdout, &stderr)
+			if err != nil {
+				stderr.WriteString(err.Error())
+			}
+			return stdout.String(), stderr.String(), err
+		}, nil
+	})
+}
